soulstep-catalog-api/requirements.txt
# Python 3.14+ (or 3.11+). Create venv with: python3 -m venv .venv && source .venv/bin/activate
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.6
sqlmodel>=0.0.14
alembic>=1.13.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pillow>=10.0.0
slowapi>=0.1.9
resend>=2.0.0
google-cloud-translate>=3.15.0
google-cloud-storage>=2.14.0
python-json-logger>=2.0.7
prometheus-fastapi-instrumentator>=6.1.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
httpx>=0.27.0
ruff>=0.3.0
